# [Merged by Bors] - Refactor the environment for runtime performance