--- a/boa/src/builtins/function/arguments.rs
+++ b/boa/src/builtins/function/arguments.rs
@@ -1,14 +1,15 @@
 use crate::{
     builtins::Array,
-    environment::lexical_environment::Environment,
     gc::{Finalize, Trace},
     object::{FunctionBuilder, JsObject, ObjectData},
-    property::PropertyDescriptor,
+    property::{PropertyDescriptor, PropertyKey},
+    realm::DeclarativeEnvironment,
     symbol::{self, WellKnownSymbols},
     syntax::ast::node::FormalParameter,
     Context, JsValue,
 };
-use rustc_hash::FxHashSet;
+use gc::Gc;
+use rustc_hash::FxHashMap;
 
 #[derive(Debug, Clone, Trace, Finalize)]
 pub struct MappedArguments(JsObject);
@@ -111,7 +112,7 @@ impl Arguments {
         func: &JsObject,
         formals: &[FormalParameter],
         arguments_list: &[JsValue],
-        env: &Environment,
+        env: &Gc<DeclarativeEnvironment>,
         context: &mut Context,
     ) -> JsObject {
         // 1. Assert: formals does not contain a rest parameter, any binding patterns, or any initializers.
@@ -160,86 +161,82 @@ impl Arguments {
         .expect("DefinePropertyOrThrow must not fail per the spec");
 
         // 17. Let mappedNames be a new empty List.
-        // using a set to optimize `contains`
-        let mut mapped_names = FxHashSet::default();
-
         // 12. Let parameterNames be the BoundNames of formals.
         // 13. Let numberOfParameters be the number of elements in parameterNames.
         // 18. Set index to numberOfParameters - 1.
         // 19. Repeat, while index ≥ 0,
         // a. Let name be parameterNames[index].
 
-        for (index, parameter_name_vec) in
-            formals.iter().map(FormalParameter::names).enumerate().rev()
-        {
-            for parameter_name in parameter_name_vec.iter().copied() {
-                // b. If name is not an element of mappedNames, then
-                if !mapped_names.contains(&parameter_name) {
-                    // i. Add name as an element of the list mappedNames.
-                    mapped_names.insert(parameter_name);
-                    // ii. If index < len, then
-                    if index < len {
-                        // 1. Let g be MakeArgGetter(name, env).
-                        // https://tc39.es/ecma262/#sec-makearggetter
-                        let g = {
-                            // 2. Let getter be ! CreateBuiltinFunction(getterClosure, 0, "", « »).
-                            // 3. NOTE: getter is never directly accessible to ECMAScript code.
-                            // 4. Return getter.
-                            FunctionBuilder::closure_with_captures(
-                                context,
-                                // 1. Let getterClosure be a new Abstract Closure with no parameters that captures
-                                // name and env and performs the following steps when called:
-                                |_, _, captures, context| {
-                                    captures.0.get_binding_value(captures.1, false, context)
-                                },
-                                (env.clone(), parameter_name),
-                            )
-                            .length(0)
-                            .name("")
-                            .build()
-                        };
-                        // 2. Let p be MakeArgSetter(name, env).
-                        // https://tc39.es/ecma262/#sec-makeargsetter
-                        let p = {
-                            // 2. Let setter be ! CreateBuiltinFunction(setterClosure, 1, "", « »).
-                            // 3. NOTE: setter is never directly accessible to ECMAScript code.
-                            // 4. Return setter.
-                            FunctionBuilder::closure_with_captures(
-                                context,
-                                // 1. Let setterClosure be a new Abstract Closure with parameters (value) that captures
-                                // name and env and performs the following steps when called:
-                                |_, args, captures, context| {
-                                    let value = args.get(0).cloned().unwrap_or_default();
-                                    // a. Return env.SetMutableBinding(name, value, false).
-                                    captures
-                                        .0
-                                        .set_mutable_binding(captures.1, value, false, context)
-                                        .map(|_| JsValue::Undefined)
-                                    // Ok(JsValue::Undefined)
-                                },
-                                (env.clone(), parameter_name),
-                            )
-                            .length(1)
-                            .name("")
-                            .build()
-                        };
-
-                        // 3. Perform map.[[DefineOwnProperty]](! ToString(𝔽(index)), PropertyDescriptor {
-                        // [[Set]]: p, [[Get]]: g, [[Enumerable]]: false, [[Configurable]]: true }).
-                        map.__define_own_property__(
-                            index.into(),
-                            PropertyDescriptor::builder()
-                                .set(p)
-                                .get(g)
-                                .enumerable(false)
-                                .configurable(true)
-                                .build(),
-                            context,
-                        )
-                        .expect("[[DefineOwnProperty]] must not fail per the spec");
-                    }
+        // Get the mapping of the environment binding index to the index of the object property.
+        // Parameters can be defined multiple times, so the mapping is not one to one.
+        let mut bindings = FxHashMap::default();
+        let mut property_index = 0;
+        'outer: for formal in formals {
+            for name in formal.names() {
+                if property_index >= len {
+                    break 'outer;
                 }
+                let binding_index = bindings.len() + 1;
+                let entry = bindings
+                    .entry(name)
+                    .or_insert((binding_index, property_index));
+                entry.1 = property_index;
+                property_index += 1;
             }
+        }
+        for (binding_index, property_index) in bindings.values() {
+            // 1. Let g be MakeArgGetter(name, env).
+            // https://tc39.es/ecma262/#sec-makearggetter
+            let g = {
+                // 2. Let getter be ! CreateBuiltinFunction(getterClosure, 0, "", « »).
+                // 3. NOTE: getter is never directly accessible to ECMAScript code.
+                // 4. Return getter.
+                FunctionBuilder::closure_with_captures(
+                    context,
+                    // 1. Let getterClosure be a new Abstract Closure with no parameters that captures
+                    // name and env and performs the following steps when called:
+                    |_, _, captures, _| Ok(captures.0.get(captures.1)),
+                    (env.clone(), *binding_index),
+                )
+                .length(0)
+                .name("")
+                .build()
+            };
+            // 2. Let p be MakeArgSetter(name, env).
+            // https://tc39.es/ecma262/#sec-makeargsetter
+            let p = {
+                // 2. Let setter be ! CreateBuiltinFunction(setterClosure, 1, "", « »).
+                // 3. NOTE: setter is never directly accessible to ECMAScript code.
+                // 4. Return setter.
+                FunctionBuilder::closure_with_captures(
+                    context,
+                    // 1. Let setterClosure be a new Abstract Closure with parameters (value) that captures
+                    // name and env and performs the following steps when called:
+                    |_, args, captures, _| {
+                        let value = args.get(0).cloned().unwrap_or_default();
+                        captures.0.set(captures.1, value);
+                        Ok(JsValue::undefined())
+                    },
+                    (env.clone(), *binding_index),
+                )
+                .length(1)
+                .name("")
+                .build()
+            };
+
+            // 3. Perform map.[[DefineOwnProperty]](! ToString(𝔽(index)), PropertyDescriptor {
+            // [[Set]]: p, [[Get]]: g, [[Enumerable]]: false, [[Configurable]]: true }).
+            map.__define_own_property__(
+                PropertyKey::from(*property_index),
+                PropertyDescriptor::builder()
+                    .set(p)
+                    .get(g)
+                    .enumerable(false)
+                    .configurable(true)
+                    .build(),
+                context,
+            )
+            .expect("[[DefineOwnProperty]] must not fail per the spec");
         }
 
         // 20. Perform ! DefinePropertyOrThrow(obj, @@iterator, PropertyDescriptor {