--- a/boa/src/builtins/function/arguments.rs
+++ b/boa/src/builtins/function/arguments.rs
@@ -58,7 +58,7 @@ impl Arguments {
                 .configurable(true),
             context,
         )
-        .expect("DefinePropertyOrThrow must not fail per the spec");
+        .expect("Defining new own properties for a new ordinary object cannot fail");
 
         // 5. Let index be 0.
         // 6. Repeat, while index < len,
@@ -66,7 +66,7 @@ impl Arguments {
             // a. Let val be argumentsList[index].
             // b. Perform ! CreateDataPropertyOrThrow(obj, ! ToString(𝔽(index)), val).
             obj.create_data_property_or_throw(index, value, context)
-                .expect("CreateDataPropertyOrThrow must not fail per the spec");
+                .expect("Defining new own properties for a new ordinary object cannot fail");
 
             // c. Set index to index + 1.
         }
@@ -83,7 +83,7 @@ impl Arguments {
                 .configurable(true),
             context,
         )
-        .expect("DefinePropertyOrThrow must not fail per the spec");
+        .expect("Defining new own properties for a new ordinary object cannot fail");
 
         let throw_type_error = context.intrinsics().throw_type_error();
 
@@ -99,7 +99,7 @@ impl Arguments {
                 .configurable(false),
             context,
         )
-        .expect("DefinePropertyOrThrow must not fail per the spec");
+        .expect("Defining new own properties for a new ordinary object cannot fail");
 
         // 9. Return obj.
         obj
@@ -143,7 +143,7 @@ impl Arguments {
             // a. Let val be argumentsList[index].
             // b. Perform ! CreateDataPropertyOrThrow(obj, ! ToString(𝔽(index)), val).
             obj.create_data_property_or_throw(index, val, context)
-                .expect("CreateDataPropertyOrThrow must not fail per the spec");
+                .expect("Defining new own properties for a new ordinary object cannot fail");
             // c. Set index to index + 1.
         }
 
@@ -158,17 +158,38 @@ impl Arguments {
                 .configurable(true),
             context,
         )
-        .expect("DefinePropertyOrThrow must not fail per the spec");
+        .expect("Defining new own properties for a new ordinary object cannot fail");
 
-        // 17. Let mappedNames be a new empty List.
-        // 12. Let parameterNames be the BoundNames of formals.
-        // 13. Let numberOfParameters be the number of elements in parameterNames.
-        // 18. Set index to numberOfParameters - 1.
-        // 19. Repeat, while index ≥ 0,
-        // a. Let name be parameterNames[index].
+        // The section 17-19 differs from the spec, due to the way the runtime environments work.
+        //
+        // This section creates getters and setters for all mapped arguments.
+        // Getting and setting values on the `arguments` object will actually access the bindings in the environment:
+        // ```
+        // function f(a) {console.log(a); arguments[0] = 1; console.log(a)};
+        // f(0) // 0, 1
+        // ```
+        //
+        // The spec assumes, that identifiers are used at runtime to reference bindings in the environment.
+        // We use indices to access environment bindings at runtime.
+        // To map to function parameters to binding indices, we use the fact, that bindings in a
+        // function environment start with all of the arguments in order:
+        // `function f (a,b,c)`
+        // | binding index | `arguments` property key | identifier |
+        // | 0             | 0                        | a          |
+        // | 1             | 1                        | b          |
+        // | 2             | 2                        | c          |
+        //
+        // Notice that the binding index does not correspond to the argument index:
+        // `function f (a,a,b)` => binding indices 0 (a), 1 (b), 2 (c)
+        // | binding index | `arguments` property key | identifier |
+        // | -             | 0                        | -          |
+        // | 0             | 1                        | a          |
+        // | 1             | 2                        | b          |
+        // While the `arguments` object contains all arguments, they must not be all bound.
+        // In the case of duplicate parameter names, the last one is bound as the environment binding.
+        //
+        // The following logic implements the steps 17-19 adjusted for our environment structure.
 
-        // Get the mapping of the environment binding index to the index of the object property.
-        // Parameters can be defined multiple times, so the mapping is not one to one.
         let mut bindings = FxHashMap::default();
         let mut property_index = 0;
         'outer: for formal in formals {
@@ -185,7 +206,7 @@ impl Arguments {
             }
         }
         for (binding_index, property_index) in bindings.values() {
-            // 1. Let g be MakeArgGetter(name, env).
+            // 19.b.ii.1. Let g be MakeArgGetter(name, env).
             // https://tc39.es/ecma262/#sec-makearggetter
             let g = {
                 // 2. Let getter be ! CreateBuiltinFunction(getterClosure, 0, "", « »).
@@ -201,7 +222,7 @@ impl Arguments {
                 .length(0)
                 .build()
             };
-            // 2. Let p be MakeArgSetter(name, env).
+            // 19.b.ii.2. Let p be MakeArgSetter(name, env).
             // https://tc39.es/ecma262/#sec-makeargsetter
             let p = {
                 // 2. Let setter be ! CreateBuiltinFunction(setterClosure, 1, "", « »).
@@ -222,7 +243,7 @@ impl Arguments {
                 .build()
             };
 
-            // 3. Perform map.[[DefineOwnProperty]](! ToString(𝔽(index)), PropertyDescriptor {
+            // 19.b.ii.3. Perform map.[[DefineOwnProperty]](! ToString(𝔽(index)), PropertyDescriptor {
             // [[Set]]: p, [[Get]]: g, [[Enumerable]]: false, [[Configurable]]: true }).
             map.__define_own_property__(
                 PropertyKey::from(*property_index),
@@ -234,7 +255,7 @@ impl Arguments {
                     .build(),
                 context,
             )
-            .expect("[[DefineOwnProperty]] must not fail per the spec");
+            .expect("Defining new own properties for a new ordinary object cannot fail");
         }
 
         // 20. Perform ! DefinePropertyOrThrow(obj, @@iterator, PropertyDescriptor {
@@ -249,7 +270,7 @@ impl Arguments {
                 .configurable(true),
             context,
         )
-        .expect("DefinePropertyOrThrow must not fail per the spec");
+        .expect("Defining new own properties for a new ordinary object cannot fail");
 
         // 21. Perform ! DefinePropertyOrThrow(obj, "callee", PropertyDescriptor {
         // [[Value]]: func, [[Writable]]: true, [[Enumerable]]: false, [[Configurable]]: true }).
@@ -262,7 +283,7 @@ impl Arguments {
                 .configurable(true),
             context,
         )
-        .expect("DefinePropertyOrThrow must not fail per the spec");
+        .expect("Defining new own properties for a new ordinary object cannot fail");
 
         // 22. Return obj.
         obj