--- a/boa/src/builtins/regexp/mod.rs
+++ b/boa/src/builtins/regexp/mod.rs
@@ -308,7 +308,7 @@ impl RegExp {
     pub(crate) fn create(p: JsValue, f: JsValue, context: &mut Context) -> JsResult<JsValue> {
         // 1. Let obj be ? RegExpAlloc(%RegExp%).
         let obj = Self::alloc(
-            &context.global_object().get(Self::NAME, context)?,
+            &context.global_object().clone().get(Self::NAME, context)?,
             &[],
             context,
         )?;