--- a/boa/src/environment/mod.rs
+++ b/boa/src/environment/mod.rs
@@ -1,3 +1,0 @@
-//! Environment handling, lexical, object, function and declarative records
-
-pub mod lexical_environment;